modularized artist dropdown
