+++ client/src/components/programs/EditProgram.jsx
@@ -21,7 +21,10 @@ import {
 
 
 import {CancelIcon} from '../../assets/CancelIcon';
-
+import {CloseFilledIcon} from '../../assets/CloseFilledIcon';
+import {RepeatIcon} from '../../assets/RepeatIcon';
+import {ClockFilledIcon} from '../../assets/ClockFilledIcon';
+import {CalendarIcon} from '../../assets/CalendarIcon';
 import {EmailIcon} from '../../assets/EmailIcon';
 import {PlusFilledIcon} from '../../assets/PlusFilledIcon';
 
@@ -34,6 +37,7 @@ import { useParams } from "react-router";
 import Navbar from "../navbar/Navbar";
 import React from 'react';
 
+import { ArtistsDropdown } from "./programComponents/ArtistsDropdown";
 import { PayeesDropdown } from "./programComponents/PayeesDropdown"
 import { LocationDropdown } from "./programComponents/LocationDropdown"
 import { RoomInformation } from "./programComponents/RoomInformation"
@@ -351,77 +355,16 @@ const payees = eventClientResponse.data
                 setSelectedDays={setSelectedDays}
               />
 
-
-
-
-
-
-              <div id="instructorContainer">
-                <div id="instructors">
-                  <div id="instructorSelection">
-                    <Box>
-                      <div id="instructorInputContainer">
-                        <Input
-                          placeholder="Instructor..."
-                          onChange={(e) => {
-                            getInstructorResults(e.target.value);
-                            setInstructorSearchTerm(e.target.value);
-                          }}
-                          value={instructorSearchTerm} id="instructorInput"/>
-                        <PlusFilledIcon />
-                      </div>
-
-                      {searchedInstructors.length > 0 && (
-                        <Box id="instructorDropdown">
-                          {searchedInstructors.map((instructor) => (
-                            <Box
-                              key={instructor.id}
-                              onClick={() => {
-                                const alreadySelected = selectedInstructors.find(
-                                  (instr) => instr.id.toString() === instructor.id
-                                );
-                                if (instructor && !alreadySelected) {
-                                  setSelectedInstructors((prevItems) => [...prevItems, instructor]);
-                                  const filteredInstructors = searchedInstructors.filter(
-                                    (instr) => instructor.id !== instr.id.toString()
-                                  );
-                                  setSearchedInstructors(filteredInstructors);
-                                }
-                              }}
-                              style={{
-                                padding: "10px",
-                                fontSize: "16px",
-                                cursor: "pointer",
-                                transition: "0.2s",
-                              }}
-                              bg="#F6F6F6"
-                              _hover={{ bg: "#D9D9D9" }}
-                            >
-                              {instructor.name}
-                            </Box>
-                          ))}
-                        </Box>
-                      )}
-                    </Box>
-                  </div>
-                </div>
-                <div id="instructorTags">
-                  {selectedInstructors.length > 0 ? (
-                    selectedInstructors.map((instructor, ind) => (
-                      <div className="instructorTag" key={ind}>
-                        <Icon fontSize="lg" onClick={() => {
-                            setSelectedInstructors(prevItems =>
-                              prevItems.filter(item => item.id !== instructor.id));
-                          }}><CloseFilledIcon /></Icon>
-                        <Tag value={instructor.id}>
-                          {instructor.name}
-                        </Tag>
-                      </div>
-                    ))
-                  ) : <div></div> }
-                </div>
-              </div>
-
+              <ArtistsDropdown 
+                instructorSearchTerm={instructorSearchTerm}
+                searchedInstructors={searchedInstructors}
+                selectedInstructors={selectedInstructors}
+                setSelectedInstructors={setSelectedInstructors}
+                setSearchedInstructors={setSearchedInstructors}
+                getInstructorResults={getInstructorResults}
+                setInstructorSearchTerm={setInstructorSearchTerm}
+              />
+              
               <PayeesDropdown
                 payeeSearchTerm={payeeSearchTerm}
                 searchedPayees={searchedPayees}
